Raw PCMFitModelMappings object from the SURFACEOU model fit to the mammal data including measurement error.

Raw PCMFitModelMappings object from the SURFACEOU model fit to the mammal data including measurement error.

fitMappings_SURFACEOU_best_clade_DataWithSEs

Format

This is a PCMFitModelMappings object. See https://venelin.github.io/PCMFit/index.html for further documentation.

Note

For additional information, read SI Appendix, Section H from Mitov et al. 2019.

References

Mitov et al. 2019: Mitov, V., Bartoszek, K., & Stadler, T. (2019). Automatic generation of evolutionary hypotheses using mixed Gaussian phylogenetic models. Proceedings of the National", Academy of Sciences of the United States of America. http://doi.org/10.1073/pnas.1813823116.

Bininda-Emonds et al. 2007: Bininda-Emonds, O. R. P., Cardillo, M., Jones, K. E., MacPhee, R. D. E., Beck, R. M. D., Grenyer, R., et al. (2007). The delayed rise of present-day mammals. Nature, 446(7135), 507–512. http://doi.org/10.1038/nature05634.

Boddy et al. 2012a: Boddy, A. M., McGowen, M. R., Sherwood, C. C., Grossman, L. I., Goodman, M., & Wildman, D. E. (2012). Comparative analysis of encephalization in mammals reveals relaxed constraints on anthropoid primate and cetacean brain scaling. Journal of Evolutionary Biology, 25(5), 981–994. http://doi.org/10.1111/j.1420-9101.2012.02491.x

Boddy et al. 2012b: Boddy AM, McGowen MR, Sherwood CC, Grossman LI, Goodman M, Wildman DE (2012) Data from: Comparative analysis of encephalization in mammals reveals relaxed constraints on anthropoid primate and cetacean brain scaling. Dryad Digital Repository. https://doi.org/10.5061/dryad.5kh0b362.